# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bzcom CRM — Customer Request Management System API
  description: >
    REST API quản lý yêu cầu khách hàng (bug / feature / inquiry) cho Bzcom.
    Mọi response dùng envelope thống nhất `{ status, message, data }`.
    Xác thực bằng JWT Bearer token (trừ login, refresh, logout và đăng ký CLIENT).
  version: 1.0.0
  contact:
    name: KITS Hanoi OJT 2026 — Bzcom Developer Team
servers:
  - url: http://localhost:8080
    description: Local (Docker Compose)
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập, cấp JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Đăng xuất phiên hiện tại (thu hồi refresh token; không gửi Bearer access token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Làm mới access token và xoay vòng refresh token (không gửi Bearer access token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: Token mới; refresh token cũ bị thu hồi
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─────────────────────────── MEMBER ───────────────────────────
...
  /api/requests:
    post:
      tags: [Request]
      summary: Tạo yêu cầu (CLIENT only). HIGH → tự sinh alert cho ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestCreateRequest' }
      responses:
        '201':
          description: Tạo request thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RequestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [Request]
      summary: Danh sách request (phân trang / sort / filter tổ hợp; phạm vi theo role)
      description: >
        ADMIN thấy tất cả; CLIENT chỉ thấy request mình tạo;
        DEVELOPER chỉ thấy request được gán cho mình.
      parameters:
        - { name: page,     in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size,     in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
        - name: sort
          in: query
          description: 'field,direction; field hợp lệ: id, title, category, priority, status, clientId, assignedDeveloperId, createdAt, updatedAt'
          schema: { type: string, example: 'createdAt,desc' }
        - { name: status,   in: query, schema: { $ref: '#/components/schemas/RequestStatus' } }
        - { name: category, in: query, schema: { $ref: '#/components/schemas/Category' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/Priority' } }
        - { name: keyword,  in: query, schema: { type: string }, description: 'tìm trong title/description' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PageRequestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/requests/{id}:
    get:
      tags: [Request]
      summary: Chi tiết một request (kiểm quyền xem theo role)
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Chi tiết request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RequestResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/requests/stats:
    get:
      tags: [Request]
      summary: Thống kê (ADMIN only) — tổng, completion rate, theo category, theo developer
      responses:
        '200':
          description: Thống kê
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/StatsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ─────────────────────────── ALERT ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Dữ liệu không hợp lệ (validation)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 400, message: 'title: must not be blank', data: null }
    Unauthorized:
      description: Chưa đăng nhập hoặc token không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 401, message: 'Unauthorized', data: null }
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 403, message: 'Forbidden', data: null }
    NotFound:
      description: Không tìm thấy resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
          example: { status: 404, message: 'Request not found: 999', data: null }
...
  schemas:
    ApiResponse:
      type: object
      required: [status, message, data]
      properties:
        status:  { type: integer, example: 200 }
        message: { type: string, example: success }
        data:    { nullable: true }
    RequestStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, DONE]
    Category:
      type: string
      enum: [BUG, FEATURE, INQUIRY]
    Priority:
      type: string
      enum: [HIGH, MEDIUM, LOW]
    Role:
      type: string
      enum: [ADMIN, DEVELOPER, CLIENT]
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:    { type: string, format: email, example: admin@bzcom.com }
        password: { type: string, maxLength: 72, example: '1234' }
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 512, description: 'opaque refresh token đã nhận khi login/refresh' }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, role]
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        tokenType:    { type: string, example: Bearer }
        role:         { $ref: '#/components/schemas/Role' }
...
    RequestCreateRequest:
      type: object
      additionalProperties: false
      required: [title, category, priority]
      properties:
        title:       { type: string, minLength: 1, maxLength: 200, example: 'Login fails' }
        description: { type: string, maxLength: 4000, example: 'When I click login, page shows 500 error' }
        category:    { $ref: '#/components/schemas/Category' }
        priority:    { $ref: '#/components/schemas/Priority' }
...
    RequestResponse:
      type: object
      properties:
        id:          { type: integer, format: int64 }
        title:       { type: string }
        description: { type: string, nullable: true }
        category:    { $ref: '#/components/schemas/Category' }
        priority:    { $ref: '#/components/schemas/Priority' }
        status:      { $ref: '#/components/schemas/RequestStatus' }
        clientId:    { type: integer, format: int64 }
        assignedDeveloperId: { type: integer, format: int64, nullable: true }
        version:     { type: integer, minimum: 0, description: 'Gửi lại ở mọi PATCH để phát hiện stale update' }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    PageRequestResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/RequestResponse' }
        page:          { type: integer, example: 0 }
        size:          { type: integer, example: 10 }
        totalElements: { type: integer, example: 42 }
        totalPages:    { type: integer, example: 5 }
...
    StatsResponse:
      type: object
      properties:
        total:          { type: integer, format: int64, example: 42 }
        completed:      { type: integer, format: int64, example: 18 }
        completionRate: { type: number, format: double, example: 0.43 }
        byCategory:
          type: object
          additionalProperties: { type: integer, format: int64 }
          example: { BUG: 20, FEATURE: 15, INQUIRY: 7 }
        byDeveloper:
          type: array
          items:
            type: object
            properties:
              developerId:   { type: integer, format: int64 }
              developerName: { type: string }
              assignedCount: { type: integer }
              doneCount:     { type: integer }
...
security:
  - bearerAuth: []